chore: separate Engine import/create flows; consistent styling
